fixed longest axis float math line problem

+++ plantcv.py
@@ -563,7 +563,7 @@ def find_objects(img, mask, device, debug=False):
 def define_roi(img, shape, device, roi=None, roi_input='default', debug=False, adjust=False, x_adj=0, y_adj=0, w_adj=0, h_adj=0, ):
   # img = img to overlay roi 
   # roi =default (None) or user input ROI image, object area should be white and background should be black, has not been optimized for more than one ROI
-  # roi_input = type of file roi_base is, either 'binary' or 'rgb'
+  # roi_input = type of file roi_base is, either 'binary', 'rgb', or 'default' (no ROI inputted)
   # shape = desired shape of final roi, either 'rectangle' or 'circle', if  user inputs rectangular roi but chooses 'circle' for shape then a circle is fitted around rectangular roi (and vice versa)
   # device = device number.  Used to count steps in the pipeline
   # debug = True/False. If True, print image
@@ -978,8 +978,8 @@ def analyze_object(img,imgname,obj, mask, device, debug=False,filename=False):
       xintercept1=0
       cv2.line(background1,(iy,caliper_mid_y),(0,caliper_mid_y),(255),1)
     else:
-      xintercept=(-b_line/slope)
-      xintercept1=(iy-b_line)/slope
+      xintercept=int(-b_line/slope)
+      xintercept1=int((iy-b_line)/slope)
       cv2.line(background1,(xintercept1,iy),(xintercept,0),(255),1)
     
     ret1,line_binary = cv2.threshold(background1, 0, 255, cv2.THRESH_BINARY)
